requirements.txt
streamlit
google-generativeai
firecrawl
pandas
python-dotenv